requirements.txt
streamlit
scikit-learn
joblib
opencv-python-headless
numpy
Pillow
matplotlib